# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_fisker_zjump-backend_internal_model.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  internal_api_handler_k8s.AddPermissionRequest:
    properties:
      action:
        description: read, write, delete, admin
        type: string
      clusterId:
        type: string
      namespace:
        type: string
      resourceName:
        description: 资源名称（可选）
        type: string
      resourceType:
        description: deployment, statefulset, service, pod, ingress, namespace
        type: string
      roleId:
        description: 角色ID（角色级权限）
        type: string
      userId:
        description: 用户ID（用户级权限）
        type: string
    required:
    - action
    - clusterId
    - namespace
    type: object
  internal_api_handler_k8s.CheckPermissionRequest:
    properties:
      action:
        description: read, write, delete, admin
        type: string
      cluster_id:
        type: string
      namespace:
        type: string
      resource_name:
        type: string
      resource_type:
        type: string
    required:
    - action
    type: object
...
  internal_api_handler_k8s.RemovePermissionRequest:
    properties:
      action:
        type: string
      clusterId:
        type: string
      namespace:
        type: string
      resourceName:
        type: string
      resourceType:
        type: string
      roleId:
        type: string
      userId:
        type: string
    required:
    - action
    - clusterId
    - namespace
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@keyops.com
    name: API Support
  description: KeyOps 基础设施管理平台 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: KeyOps API
  version: "2.0"
paths:
...
  /api/k8s/clusters/:id/permissions:
    get:
      consumes:
      - application/json
      parameters:
      - description: 集群ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      security:
      - BearerAuth: []
      summary: 获取集群的所有权限
      tags:
      - K8s Permission
...
  /api/k8s/permissions:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除权限请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler_k8s.RemovePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      security:
      - BearerAuth: []
      summary: 删除K8s资源权限
      tags:
      - K8s Permission
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: query
        name: userId
        type: string
      - description: 角色ID
        in: query
        name: roleId
        type: string
      - description: 集群ID
        in: query
        name: clusterId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      security:
      - BearerAuth: []
      summary: 获取K8s资源权限列表
      tags:
      - K8s Permission
    post:
      consumes:
      - application/json
      parameters:
      - description: 添加权限请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler_k8s.AddPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      security:
      - BearerAuth: []
      summary: 添加K8s资源权限
      tags:
      - K8s Permission
  /api/k8s/permissions/check:
    post:
      consumes:
      - application/json
      parameters:
      - description: 检查权限请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler_k8s.CheckPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      security:
      - BearerAuth: []
      summary: 检查K8s资源权限
      tags:
      - K8s Permission
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"